kprobe

Trait KprobeAuxiliaryOps

pub trait KprobeAuxiliaryOps: Send + Debug {
    // Required methods
    fn copy_memory(
        src: *const u8,
        dst: *mut u8,
        len: usize,
        user_pid: Option<i32>,
    );
    fn set_writeable_for_address<F: FnOnce(*mut u8)>(
        address: usize,
        len: usize,
        user_pid: Option<i32>,
        action: F,
    );
    fn alloc_kernel_exec_memory() -> *mut u8;
    fn free_kernel_exec_memory(ptr: *mut u8);
    fn alloc_user_exec_memory<F: FnOnce(*mut u8)>(
        pid: Option<i32>,
        action: F,
    ) -> *mut u8;
    fn free_user_exec_memory(pid: Option<i32>, ptr: *mut u8);
    fn insert_kretprobe_instance_to_task(instance: RetprobeInstance);
    fn pop_kretprobe_instance_from_task() -> RetprobeInstance;
}

The auxiliary operations required for kprobes.

Required Methods

fn copy_memory(src: *const u8, dst: *mut u8, len: usize, user_pid: Option<i32>)

Copy memory from source to destination. If user_pid is Some(pid), it indicates that the src is in user space. The dst is always in kernel space.

fn set_writeable_for_address<F: FnOnce(*mut u8)>( address: usize, len: usize, user_pid: Option<i32>, action: F, )

Enable or disable write permission for the specified address.

fn alloc_kernel_exec_memory() -> *mut u8

Allocate executable memory(one page)

fn free_kernel_exec_memory(ptr: *mut u8)

Deallocate executable memory(one page)

fn alloc_user_exec_memory<F: FnOnce(*mut u8)>( pid: Option<i32>, action: F, ) -> *mut u8

Allocate user executable memory(one page)

fn free_user_exec_memory(pid: Option<i32>, ptr: *mut u8)

Deallocate user executable memory(one page)

fn insert_kretprobe_instance_to_task(instance: RetprobeInstance)

Insert a kretprobe instance to the current task

fn pop_kretprobe_instance_from_task() -> RetprobeInstance

Pop a kretprobe instance from the current task

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.