kprobe

Trait KprobeOps

pub trait KprobeOps: Send {
    // Required methods
    fn return_address(&self) -> usize;
    fn single_step_address(&self) -> usize;
    fn debug_address(&self) -> usize;
    fn break_address(&self) -> usize;
}

The operations available for kprobes.

Required Methods

fn return_address(&self) -> usize

The address of the instruction that program should return to

fn single_step_address(&self) -> usize

The address of the instruction that saved the original instruction

Usually, the original instruction at the probe point is saved in an array. Depending on the architecture, necessary instructions may be filled in after the saved instruction. For example, x86 architecture supports single-step execution, while other architectures usually do not. Therefore, we use the break exception to simulate it, so a breakpoint instruction will be filled in.

fn debug_address(&self) -> usize

The address of the instruction that caused the single step exception

fn break_address(&self) -> usize

The address of the instruction that caused the break exception

It is usually equal to the address of the instruction that used to set the probe point.